Rudimentary generation of VBO / IBO for faces.

+++ src/Modules/Visualization/ShowField.h
@@ -101,7 +101,7 @@ namespace Visualization {
         const std::vector<Core::Geometry::Point>  &points,
         const std::vector<Core::Geometry::Vector> &normals,
         bool withNormals,
-        int64_t& iboBufferIndex,
+        uint32_t& iboBufferIndex,
         CPM_VAR_BUFFER_NS::VarBuffer* iboBuffer,
         CPM_VAR_BUFFER_NS::VarBuffer* vboBuffer,
         Core::Datatypes::GeometryObject::ColorScheme colorScheme,
